fix: move permissions and filters file out of utils folder and make a helper method to reset password
what: code improvement

why: code refactoring

impact: LOW
